Lau - DEC - Versión Final

+++ trunk/Promethee/Promethee/Scripts/UIUtility.cs
@@ -30,16 +30,15 @@ namespace Promethee.Scripts
             if (Directory.Exists(ImgPath))
             {
                 foreach (string item in Directory.GetFiles(ImgPath, archivo, SearchOption.TopDirectoryOnly))
-                {
                     File.Delete(item);
-                }
                 if (HttpContext.Current != null)
                 {
                     ImgPath = "~/Images/TMP";
-                    string miruta = HttpContext.Current.Server.MapPath(ImgPath).ToString();
-                    foreach (string item in Directory.GetFiles(miruta, archivo, SearchOption.TopDirectoryOnly))
+                    if (Directory.Exists(HttpContext.Current.Server.MapPath(ImgPath)))
                     {
-                        File.Delete(item);
+                        string miruta = HttpContext.Current.Server.MapPath(ImgPath).ToString();
+                        foreach (string item in Directory.GetFiles(miruta, archivo, SearchOption.TopDirectoryOnly))
+                            File.Delete(item);
                     }
                 }
             }
